Migrate to asynchronous dns lookups
The underlying synchronous trust dns resolver will spawn a new tokio
runtime to block itself, as under the covers it uses an asynchronous
resolver. One can't nest runtimes like this (dness wouldn't start). The
first is to use the asynchronous resolver directly. This commit also
adds a test to make sure the default behavior when executing dness with
no arguments successfully resolves to a WAN ip (instead of erroring).
This test should not be executed on CI when the cpu that is compiling
the target differs from the target.