Introduce alternative WAN ip resolvers
Previously it wasn't possible to configure dness to use an alternative
ip resolver other than opendns. This has worked fine for the most part,
but some users are reporting that the opendns method doesn't work on
their networks. The fix is to introduce configuration so that the user
can pick amongst ip resolvers. The default will continue to be opendns,
but now ipify is available. One can opt into using ipify by adding the
following to the top of their config

```toml
ip_resolver = "ipify"
```