Introduce handlebars config templating
Currently the dness config contains sensitive fields that one would not
want to be public knowledge. In production the config can be guarded by
file ACLs (ie: one can have the config file owned by root.dness and make
the `dness` group read only and not world readable). This works and is
fairly standard but can be suboptimal.

One dness is not distributed this way (the service is executed as root),
and while this can be fixed I'd much rather allow the config to be world
readable and substitute in sensitive values via environment variables

This will help make dness trivially systemd sandboxable as right now
allowing a service access to root owned file is a brand new feature:
https://github.com/systemd/systemd/issues/16060