Disable live IPv6 tests by default
Although the GitHub runners have IPv6 support, they have no route to the
global internet, causing IPv6 address resolution to fail. We don't want
this to cause CI to fail, so we gate these tests behind the #[ignored]
attribute. The tests can still be run with

    cargo test -- --ignored

or

    cargo test -- --include-ignored

See actions/runner-images#668