Add initial IPv6 support for Cloudflare (#504)
* Add IP address type enum

* Also add ip_types field with new enum to CloudflareConfig

* Add tests for ip_types field

* Add IPv6 cloudflare support

* Use IpAddr instead of Ipv4Addr as the address type
* Add AAAA record handling to Cloudflare; panic on IPv6 for all other
  providers
* Add IPv6 handling to both resolvers
* Query required address types for each domain before getting IPs
* Call update_provider for each required address type (might duplicate
  work, e.g. fetching zone ID)

* resolve IP with empty config

* Contact resolver using same IP type

The resolver only knows the address that we contact it over, so to get
our IPv6 adress we need to contact it over IPv6

* handle returned IPv6 addresses

* Add tests for Cloudflare IPv6

* default to IPv4 only

* Continue when an address fails to resolve

In the case that we request both IPv4 and IPv6 addresses but are only
able to resolve one, we should continue with updating that address
intead of immediately exiting. Thanks to the fact we loop over the
successful addresses we don't need special error handling.

* Don't panic on unsupported IP type

Although we currently only pass in IPs with supported types, we don't
want to crash the program if that happens in the future.

* Error on IP resolution failure

If no IP addresses could be resolved then we don't try and update any
DNS records. I didn't realize that with the original logic this means
the failure flag wasn't set and we exit with a success code.

* Disable live IPv6 tests by default

Although the GitHub runners have IPv6 support, they have no route to the
global internet, causing IPv6 address resolution to fail. We don't want
this to cause CI to fail, so we gate these tests behind the #[ignored]
attribute. The tests can still be run with

    cargo test -- --ignored

or

    cargo test -- --include-ignored

See actions/runner-images#668

* Change default address lookup to IPv4 only

Prevents CI failure on the IPv4-only GitHub Actions runners. Also keeps
the no-config behavior the same as before.

* Initial GoDaddy IPv6 support

* Dynu IPv6 support

* Hurricane Electric IPv6 support

* No-Ip IPv6 support

* Initial Porkbun IPv6 support

* Update documentation for ip_types

* Minor formatting